Added ControlsOverviewPage.razor Added MinimapControls.razor Disabled logging by default Link target port and soource port have an internal setter now. Made zoom a double again Added fit to screen feature

+++ Blazored.Diagrams.Test/Diagrams/DiagramTests.cs
@@ -26,7 +26,7 @@ public class DiagramTests
     public void Test_Set_Zoom()
     {
         //Arrange
-        const decimal zoom = 2;
+        const double zoom = 2;
         var diagram = Instance;
 
         //Act
