+++ Blazored.Diagrams/Options/Behaviours/ZoomBehaviourOptions.cs
@@ -12,31 +12,31 @@ public class ZoomBehaviourOptions : BaseBehaviourOptions
     /// <summary>
     ///     Default value for  the zoom step.
     /// </summary>
-    [JsonIgnore] private const decimal DefaultZoomStep = 0.1m;
+    [JsonIgnore] private const double DefaultZoomStep = 0.1;
 
     /// <summary>
     ///     Default value for the minimum  zoom value.
     ///     Must be positive.
     /// </summary>
-    [JsonIgnore] private const decimal DefaultMinimumZoom = 0.1m;
+    [JsonIgnore] private const double DefaultMinimumZoom = 0.1;
 
     /// <summary>
     ///     Default value for the maximum zoom value.
     /// </summary>
-    [JsonIgnore] private const decimal DefaultMaximumZoom = 3;
+    [JsonIgnore] private const double DefaultMaximumZoom = 3;
 
     /// <summary>
     ///     Minimum allowed zoom value of the diagram.
     /// </summary>
-    public decimal MinZoom { get; set; } = DefaultMinimumZoom;
+    public double MinZoom { get; set; } = DefaultMinimumZoom;
 
     /// <summary>
     ///     Maximum allowed zoom value of the diagram.
     /// </summary>
-    public decimal MaxZoom { get; set; } = DefaultMaximumZoom;
+    public double MaxZoom { get; set; } = DefaultMaximumZoom;
 
     /// <summary>
     ///     Step value from zooming with the pointer wheel.
     /// </summary>
-    public decimal ZoomStep { get; set; } = DefaultZoomStep;
+    public double ZoomStep { get; set; } = DefaultZoomStep;
 }
